% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Run latent variable Bayesian model}
\usage{
inference(
  loaded_data,
  prior = 0.1,
  parallel = TRUE,
  n_cores = 2,
  K = 10000,
  burn_in = 1000,
  thin = 5
)
}
\arguments{
\item{loaded_data}{list of dataframes returned by \code{\link{load}}.}

\item{prior}{a numeric value indicating how much weight to assign on trascriptomic prior information.}

\item{parallel}{a boolean value to enable parallel computation.}

\item{n_cores}{number of cores to be used during parallel computation.}

\item{K}{number of MCMC iterations. # modificabile dall'utente?}

\item{burn_in}{number of initial iterations to discard. # modificabile dall'utente?}

\item{thin}{thinning value to apply to the final MCMC chain. # modificabile dall'utente?}
}
\value{
A list of two dataframe: 'isoform_results' and "normalized_isoform_results" (results normalized by gene).
}
\description{
\code{inference} run latent variable Bayesian model
}
\seealso{
links
}
\author{
name
}
