% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_data}
\alias{load_data}
\title{Load data}
\usage{
load_data(
  path_to_peptides_psm,
  path_to_peptides_intensities = NULL,
  tpm_path = NULL,
  path_fasta = NULL,
  input_type = "metamorpheus",
  abundance_type = "psm",
  PEP = FALSE,
  FDR_thd = 0.01
)
}
\arguments{
\item{path_to_peptides_psm}{a string indicating the path to psmtsv file from metamorpheus or the xml file from OpenMS.}

\item{path_to_peptides_intensities}{a string indicating the path to psmtsv file from metamorpheus with intensity values.}

\item{tpm_path}{a string indicating the path to tpm file.}

\item{path_fasta}{a string indicating the path to fasta file.}

\item{input_type}{a character string indicating the tool that outputs the peptides file: 'metamorpheus' or 'openMS'.}

\item{abundance_type}{a character string indicating the type of input: 'psm' or 'intensities'.}

\item{PEP}{boolean value indicating if Peptite Error Probability shoud be used.}

\item{FDR_thd}{numeric value indicating the False Discovery Rate filter.}
}
\value{
A list with a peptide dataframe, a peptide dataframe of unique peptides (only for PEP=TRUE) and a protein dataframe.
}
\description{
\code{load_data} read and process input data (peptides counts from metamorpheus or OpenMS, peptides intensities, tpm)
}
\seealso{
links
}
\author{
name
}
