% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_abundances.R
\name{plot_relative_abundances}
\alias{plot_relative_abundances}
\title{Plot isoform results}
\usage{
plot_relative_abundances(
  res_inference,
  gene_id,
  plot_CI = TRUE,
  normalize_gene = TRUE
)
}
\arguments{
\item{res_inference}{\code{list} of \code{data.frame} returned by \code{\link{inference}}.}

\item{gene_id}{a character string indicating the gene to be plot.}

\item{plot_CI}{logical; if TRUE (default), plot Credibility Intervals for each isoform.}

\item{normalize_gene}{logical; if TRUE (default), plot isoforms relative abundances normalized within the specified gene.}
}
\value{
A plot showing isoforms relative abundances for a specific gene.
}
\description{
\code{plot_relative_abundances} plots protein isoforms results for a specific gene together with transcripts abundances if available.
}
\examples{
# Load internal data to the package:
data_dir = system.file("extdata", package = "SIMBA")

# Define the path to the AllPeptides.psmtsv file returned by MetaMorpheus tool
path_to_peptides_psm = paste0(data_dir, "/AllPeptides.psmtsv")

# Define the path to the jurkat_isoform_kallisto.tsv with mRNA relative abundance
tpm_path = paste0(data_dir, "/jurkat_isoform_kallisto.tsv")

# Load the data
data_loaded = load_data(
  path_to_peptides_psm = path_to_peptides_psm,
  path_to_tpm = tpm_path
)

# Define the path to the map_iso_gene.csv file
path_to_map_iso_gene = paste0(data_dir, "/map_iso_gene.csv")

# Run the algorithm
set.seed(169612)
results = inference(data_loaded, prior = 0.1, map_iso_gene = path_to_map_iso_gene)

# Plotting results
plot_relative_abundances(results, gene_id = "HLA")

# For more examples see the vignettes:
#browseVignettes("SIMBA")

}
\seealso{
\code{\link{load_data}}, \code{\link{inference}}
}
\author{
Simone Tiberi \email{simone.tiberi@unibo.it} and Jordy Bollon \email{jordy.bollon@iit.it}
}
