% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_traceplot.R
\name{plot_traceplot}
\alias{plot_traceplot}
\title{Traceplot of the (thinned) posterior chain of the relative abundance of each protein isoform (i.e., pi).}
\usage{
plot_traceplot(results, protein_id)
}
\arguments{
\item{results}{a \code{list} of \code{\linkS4class{data.frame}} objects, 
computed via \code{\link{inference}}.}

\item{protein_id}{a character, indicating the protein isoform to plot.}
}
\value{
A \code{gtable} object.
}
\description{
\code{plot_traceplot} plots the traceplot of the (thinned) posterior chain 
of the relative abundance of each protein isoform (i.e., pi).
The vertical grey dashed line indicates the burn-in 
(the iterations on the left side of the burn-in are discarded in posterior analyses).
}
\examples{
# see the example of DifferentialRegulation function:
help(inference)

}
\seealso{
\code{\link{DifferentialRegulation}}
}
\author{
Simone Tiberi \email{simone.tiberi@unibo.it}
}
