% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R
\name{inference}
\alias{inference}
\title{Run latent variable Bayesian model}
\usage{
inference(
  loaded_data,
  map_iso_gene = "",
  n_cores = 1,
  K = 2000,
  burn_in = 1000,
  thin = 1
)
}
\arguments{
\item{loaded_data}{\code{list} of \code{data.frame} returned by \code{\link{load_data}}.}

\item{map_iso_gene}{a character string indicating the path to a csv file with two fields: isoform and gene name. Required to return protein
isoforms relative abundances normalized within each gene and to plot results via \code{\link{plot_relative_abundances}}.}

\item{n_cores}{the number of cores to use during algorithm execution. Default is 1.}

\item{K}{the number of MCMC iterations. Default is 2000.}

\item{burn_in}{the number of initial iterations to discard. Default is 1000.}

\item{thin}{thinning value to apply to the final MCMC chain. Default is 1.}
}
\value{
A \code{list} of two \code{data.frame}: 'isoform_results' and 'normalized_isoform_results' (relative abundances normalized
within each gene, if `map_iso_gene` is provided).
}
\description{
\code{inference} run latent variable Bayesian model taking as input the data created by \code{\link{load_data}}.
}
\examples{
# Load internal data to the package:
data_dir = system.file("extdata", package = "SIMBA")

# Define the path to the AllPeptides.psmtsv file returned by MetaMorpheus tool
path_to_peptides_psm = paste0(data_dir, "/AllPeptides.psmtsv")

# Define the path to the jurkat_isoform_kallisto.tsv with mRNA relative abundance
tpm_path = paste0(data_dir, "/jurkat_isoform_kallisto.tsv")

# Load the data
data_loaded = load_data(
  path_to_peptides_psm = path_to_peptides_psm,
  path_to_tpm = tpm_path
)

# Define the path to the map_iso_gene.csv file
path_to_map_iso_gene = paste0(data_dir, "/map_iso_gene.csv")

# Run the algorithm
set.seed(169612)
results = inference(data_loaded, prior = 0.1, map_iso_gene = path_to_map_iso_gene)
# For more examples see the vignettes:
#browseVignettes("SIMBA")

}
\seealso{
\code{\link{load_data}} and \code{\link{plot_relative_abundances}}
}
\author{
Simone Tiberi \email{simone.tiberi@unibo.it} and Jordy Bollon \email{jordy.bollon@iit.it}
}
