% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_abundances.R
\name{plot_relative_abundances}
\alias{plot_relative_abundances}
\title{Plot isoform results}
\usage{
plot_relative_abundances(x, gene_id, plot_CI = TRUE, normalize_gene = TRUE)
}
\arguments{
\item{x}{list of dataframes returned by \code{\link{inference}}.}

\item{gene_id}{a character string indicating the gene to be plot.}

\item{plot_CI}{a boolean value to plot Credibility Interval.}

\item{normalize_gene}{a boolean value to plot results normalized by gene.}
}
\value{
A plot with relative frequencies of isoforms detected for a specific gene.
}
\description{
\code{plot_relative_abundances} plot isoform results for a specific gene.
}
\seealso{
links
}
\author{
name
}
