% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_abundances.R
\name{plot_relative_abundances}
\alias{plot_relative_abundances}
\title{Plot isoform results}
\usage{
plot_relative_abundances(
  res_inference,
  gene_id,
  plot_CI = TRUE,
  normalize_gene = TRUE
)
}
\arguments{
\item{res_inference}{\code{list} of two \code{data.frame} objects
returned by \code{\link{inference}}.}

\item{gene_id}{a character string indicating the gene to be plotted.}

\item{plot_CI}{logical; if TRUE (default),
plot 0.95 level Credibility Intervals for each isoform.}

\item{normalize_gene}{logical; if TRUE (default),
plot isoform relative abundances,
normalized within the specified gene (they add to 1 within a gene).}
}
\value{
A \code{ggplot} object, showing isoform relative abundances for a specific gene.
}
\description{
\code{plot_relative_abundances} plots protein isoforms results,
obtained by \code{\link{inference}},
for a specific gene, together with transcripts abundances if available.
}
\examples{
# Load internal data to the package:
data_dir = system.file("extdata", package = "IsoBayes")

# Define the path to the AllPeptides.psmtsv file returned by MetaMorpheus tool
path_to_peptides_psm = paste0(data_dir, "/AllPeptides.psmtsv")

# Generate a SummarizedExperiment object
SE = generate_SE(path_to_peptides_psm = path_to_peptides_psm,
                 abundance_type = "psm",
                 input_type = "metamorpheus"
                 )
# Define the path to the jurkat_isoform_kallisto.tsv with mRNA relative abundance
tpm_path = paste0(data_dir, "/jurkat_isoform_kallisto.tsv")
           
# Load and process SE object
data_loaded = input_data(SE, path_to_tpm = tpm_path)

# Define the path to the map_iso_gene.csv file
path_to_map_iso_gene = paste0(data_dir, "/map_iso_gene.csv")

# Run the algorithm
set.seed(169612)
results = inference(data_loaded, map_iso_gene = path_to_map_iso_gene)

# results is a list of 3 data.frames:
names(results)

# main results:
head(results$isoform_results)

# gene abundance
head(results$gene_abundance)

# results normalized within genes (total abundance of each gene),
# useful to study alternative splicing within genes:
head(results$normalized_isoform_results)

# Plotting results, normalizing within genes
# (relative abundances add to 1 within each gene):
plot_relative_abundances(results,
    gene_id = "TUBB",
    normalize_gene = TRUE
)

# Plotting results, NOT normalized
# (relative abundances add to 1 across all isoforms in the dataset):
plot_relative_abundances(results,
    gene_id = "TUBB",
    normalize_gene = FALSE
)

# For more examples see the vignettes:
# browseVignettes("IsoBayes")

}
\seealso{
\code{\link{inference}}
}
\author{
Jordy Bollon \email{jordy.bollon@iit.it}
and Simone Tiberi \email{simone.tiberi@unibo.it}
}
