\name{IsoBayes-package}
\alias{IsoBayes-package}
\alias{IsoBayes}
\docType{package}
\title{
\packageTitle{IsoBayes}
}
\description{
\packageDescription{IsoBayes}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{IsoBayes}
Questions relative to IsoBayes should be reported as a new issue at \href{https://github.com/SimoneTiberi/IsoBayes/issues}{https://github.com/SimoneTiberi/IsoBayes/issues}.

To access the vignettes, type: \code{browseVignettes("IsoBayes")}.

\packageIndices{IsoBayes}
}
\author{
Jordy Bollon \email{jordy.bollon@iit.it},
Simone Tiberi \email{simone.tiberi@unibo.it}
}
\keyword{ package }