\name{SIMBA-package}
\alias{SIMBA-package}
\alias{SIMBA}
\docType{package}
\title{
\packageTitle{SIMBA}
}
\description{
\packageDescription{SIMBA}
}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{SIMBA}
Questions relative to SIMBA should be reported as a new issue at \href{https://github.com/SimoneTiberi/SIMBA/issues}{https://github.com/SimoneTiberi/DifferentialRegulatio/issues}.

To access the vignettes, type: \code{browseVignettes("SIMBA")}.

\packageIndices{SIMBA}
}
\author{
\packageAuthor{SIMBA}
Maintainer: \packageMaintainer{SIMBA}
}
\keyword{ package }